+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Telemonitor Changelog
 
+[**Unreleased**](https://github.com/maximilionus/Telemonitor)
+- **Sys Info** button message: Move `Architecture` to `System`
+- Pack all `TM_ControlInlineKB` buttons into enum
+
 [**0.0.2**](https://github.com/maximilionus/Telemonitor/releases/tag/v0.0.2) : 2020-06-26
 - Added whitelist check for buttons callback handler
 - Added support for easy `poetry run ...` command